% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JMAK-package.R
\docType{package}
\name{JMAK-package}
\alias{JMAK}
\alias{JMAK-package}
\title{JMAK: Johnson-Mehl-Avrami-Kolmogorov (JMAK / Avrami) kinetics tools}
\description{
Tools to compute, fit and visualize JMAK/Avrami kinetics models (isothermal nucleation+growth).
}
\author{
\strong{Maintainer}: EBA NGOLONG Jeanne Chantal \email{jeanne.eba@facsciences-uy1.cm}

}
\keyword{internal}
