% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-polymere_cristallisation.R
\docType{data}
\name{polymere_cristallisation}
\alias{polymere_cristallisation}
\title{Dataset: polymer_crystallization}
\format{
A data.frame with 60 rows and 2 variables:
\describe{
\item{t}{Time (arbitrary units).}
\item{Y}{Transformed fraction (values between 0 and 1).}
}
}
\source{
Generated by \code{data-raw/make_data.R} (simulation).
}
\usage{
polymere_cristallisation
}
\description{
Simulated dataset representing the crystallized fraction of a polymer
as a function of time. This dataset is provided for examples and tests in the package.
}
\keyword{datasets}
